backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
httpx>=0.28.0,<1.0.0
pydantic-settings>=2.7.0,<3.0.0
tomli>=2.0.0,<3.0.0
pypdf>=6.0.0,<7.0.0
python-dotenv>=1.0.1,<2.0.0
sqlalchemy[asyncio]>=2.0.36,<3.0.0
psycopg[binary]>=3.2.0,<4.0.0
alembic>=1.14.0,<2.0.0
